Wraps the commands in functions and moves out of the App component
