+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export { default as getAsteroidsCmd } from './asteroids';
+export { default as getCVCmd } from './cv';
+export { default as getGithubCmd } from './github';
+export { default as getLinkedInCmd } from './linkedin';
+export { default as getTwitterCmd } from './twitter';
