--- src/components/App/commands/index.ts
@@ -1,5 +1,0 @@
-export { default as getAsteroidsCmd } from './asteroids';
-export { default as cvCommand } from './cv';
-export { default as githubCommand } from './github';
-export { default as linkedinCommand } from './linkedin';
-export { default as twitterCommand } from './twitter';
