+++ src/descriptions/asteroids.ts
